Se esta depurando el handler del curso para mostrar cursos en la vista

+++ PI Bases y IS Grupo 1/PI Bases y IS Grupo 1/Models/Cursos.cs
@@ -24,7 +24,7 @@ namespace PIBasesISGrupo1.Models
         [Display(Name = "Ingrese su correo")]
         [RegularExpression(@"^[_a-z0-9-]+(.[_a-z0-9-]+)*@[a-z0-9-]+(.[a-z0-9-]+)*(.[a-z]{2,4})$", ErrorMessage = "Por favor ingrese un correo valido")]
         public string estado { get; set; }
-        public float precio {get; set;} 
+        public double precio {get; set;} 
         public string emailDelQueLoPropone { get; set; }
         public string emailDelEducador { get; set; }
         public string emailDelMiembroDeNucleo { get; set; }
